backend/pyproject.toml
[project]
name = "documind_backend"
version = "0.1.0"
description = "Documind AI backend"
authors = [
    { name = "ignius299792458", email = "bogati.mahesh.299.792.458@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.12,<3.14"

dependencies = [
    "fastapi>=0.135.1,<0.136.0",
    "uvicorn>=0.41.0,<0.42.0",
    "langchain>=1.2.0,<1.2.10",
    "langchain-core>=1.2.0,<2.0.0",
    "langchain-community>=0.4.1,<0.5.0",
    "langgraph>=1.0.7,<1.1.0",
    "chromadb>=1.5.5,<2.0.0",
    "tiktoken>=0.12.0,<0.13.0",
    "langchain-text-splitters>=1.1.1,<2.0.0",
    "rank-bm25>=0.2.2,<0.3.0",
    "python-dotenv>=1.2.2,<2.0.0",
    "pydantic-settings>=2.13.1,<3.0.0",
    "httpx>=0.28.1,<0.29.0",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "requests (>=2.32.5,<3.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "flashrank (>=0.2.10,<0.3.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "langchain-classic (>=1.0.2,<2.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "multipart (>=1.3.1,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]

[tool.poetry]
packages = [
    { include = "documind_backend", from = "src" }
]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "black>=26.3.0,<27.0.0",
    "ruff>=0.15.5,<0.16.0",
    "mypy>=1.19.1,<2.0.0",
    "ipython>=9.11.0,<10.0.0",
    "taskipy>=1.14.1,<2.0.0"
]

[tool.taskipy.tasks]
dev = "uvicorn documind_backend.main:app --reload"
test = "pytest"
lint = "ruff check ."
format = "black ."
typecheck = "mypy ."
check = "task format && task lint && task typecheck"

[tool.black]
line-length = 100
target-version = ['py312']

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"